streamlit_app/requirements.txt
streamlit==1.31.0
mysql-connector-python==8.2.0
pandas==2.1.4
requests==2.31.0
python-dotenv==1.0.0
plotly==5.18.0
